perf!: store source spans as u32 to shrink LabeledSpan (#180)
* perf!: store source spans as u32 to shrink LabeledSpan

Change the internal storage and public API of `SourceSpan`/`SourceOffset`
(and therefore `LabeledSpan`) from `usize` to `u32`. This shrinks
`LabeledSpan` from 48 to 40 bytes.

Source files are bounded well below 4 GiB in practice, so a 32-bit byte
offset is sufficient. `ByteOffset` is now `u32`, getters return `u32`,
and the `(usize, usize)` / `Range<usize>` convenience conversions are
replaced with their `u32` equivalents.

BREAKING CHANGE: `SourceSpan`/`SourceOffset`/`LabeledSpan` offsets and
lengths are now `u32` instead of `usize`; offsets above `u32::MAX` are no
longer representable.

* fix(clippy): use unwrap_or now that span_offset is a local

* fix: use Range::len to avoid underflow on reversed ranges

+++ src/handlers/graphical.rs
@@ -1252,7 +1252,7 @@ impl GraphicalReportHandler {
         let context = std::str::from_utf8(context_data.data()).expect("Bad utf8 detected");
         let mut line = context_data.line();
         let mut column = context_data.column();
-        let mut offset = context_data.span().offset();
+        let mut offset = context_data.span().offset() as usize;
         let mut line_offset = offset;
         let mut line_str = String::with_capacity(context.len());
         let mut lines = Vec::with_capacity(1);
@@ -1429,10 +1429,10 @@ impl FancySpan {
     }
 
     fn offset(&self) -> usize {
-        self.span.offset()
+        self.span.offset() as usize
     }
 
     fn len(&self) -> usize {
-        self.span.len()
+        self.span.len() as usize
     }
 }
